+++ Code/LYRUIParticipantPickerController.h
@@ -20,17 +20,26 @@
 @protocol LYRUIParticipantPickerControllerDelegate <NSObject>
 
 /**
- @abstract Tells the receiver that the participant selection view was dismissed without making a selection.
- @param participantSelectionViewController The participant selection view that was dismissed.
+ @abstract Tells the receiver that the participant picker was dismissed without making a selection.
+ @param participantPickerController The participant picker that was dismissed.
  */
-- (void)participantSelectionViewControllerDidCancel:(LYRUIParticipantPickerController *)participantSelectionViewController;
+- (void)participantPickerControllerDidCancel:(LYRUIParticipantPickerController *)participantPickerController;
 
 /**
- @abstract Tells the receiver that the user has selected a set of participants from a participant selection view.
- @param participantSelectionViewController The participant selection view in which the selection was made.
- @param participants The set of participants that was selected.
+ @abstract Tells the receiver that the user has selected a participant from a participant picker.
+ @param participantPickerController The participant picker in which the selection was made.
+ @param participant The participant who was selected.
  */
-- (void)participantSelectionViewController:(LYRUIParticipantPickerController *)participantSelectionViewController didSelectParticipant:(id<LYRUIParticipant>)participant;
+- (void)participantPickerController:(LYRUIParticipantPickerController *)participantPickerController didSelectParticipant:(id<LYRUIParticipant>)participant;
+
+@optional
+
+/**
+ @abstract Tells the receiver that the user has deselected a participant from a participant picker.
+ @param participantPickerController The participant picker in which the deselection was made.
+ @param participant The participant who was deselected.
+ */
+- (void)participantPickerController:(LYRUIParticipantPickerController *)participantPickerController didDeselectParticipant:(id<LYRUIParticipant>)participant;
 
 @end
 
@@ -41,22 +50,22 @@
 
 /**
  @abstract The set of participants to be presented in the picker. Each object in the returned collection must conform to the `LYRUIParticipant` protocol.
- @discussion The picker presents the returned participants in alphabetical order sectioned by the value returned by the `sectionText` property.
+ @discussion The picker presents the returned participants in alphabetical order according to `sortType` and sectioned by the corresponding first initial.
  */
-@property (nonatomic) NSSet *participants;
+- (NSSet *)participantsForParticipantPickerController:(LYRUIParticipantPickerController *)participantPickerController;
 
 /**
  @abstract Asynchronously searches for participants that match the given search text.
  @discussion Invoked by the participant picker controller when the user inputs text into the search bar. The receiver is
  to perform the search, build a set of matching participants, and then call the completion block. The controller will section
- the participants using the value returned by the `sectionText` property and present them in alphabetical order.
+ the participants and present them in alphabetical order according to the value returned by the `sortType` property.
  */
-- (void)searchForParticipantsMatchingText:(NSString *)searchText completion:(void (^)(NSSet *participants))completion;
+- (void)participantPickerController:(LYRUIParticipantPickerController *)participantPickerController searchForParticipantsMatchingText:(NSString *)searchText completion:(void (^)(NSSet *participants))completion;
 
 @end
 
 /**
- @abstract Displays a list of participants in a navigation controller and allows for searching of participants
+ @abstract Displays a list of participants in a navigation controller and allows for searching of participants.
  */
 @interface LYRUIParticipantPickerController : UINavigationController
 
@@ -66,9 +75,9 @@
 
 /**
  @abstract Creates and returns a participant picker initialized with the given set of participants.
- @param participants The set of participants to display in the picker. Each object in the given set must conform to the `LYRUIParticipant` protocol.
- @returns A new participant picker initialized with the given set of participants.
- @raises NSInvalidArgumentException Raised if any object in the given set of participants does not conform to the `LYRUIParticipant` protocol.
+ @param dataSource The object to acts as the picker's data source.
+ @param sortType The sort type to use for the participants.
+ @returns A new participant picker initialized with the given data source and sort type.
  */
 + (instancetype)participantPickerWithDataSource:(id<LYRUIParticipantPickerDataSource>)dataSource sortType:(LYRUIParticipantPickerSortType)sortType;
 
@@ -95,8 +104,9 @@
 ///---------------------------------
 
 /**
- @abstract A Boolean value that determines whether multiple participants can be selected at once.
+ @abstract A boolean value that determines whether multiple participants can be selected at once.
  @discussion The defauly value of this property is `YES`.
+ @raises NSInternalInconsistencyException Raised if the value is mutated after the receiver has been presented.
  */
 @property (nonatomic, assign) BOOL allowsMultipleSelection;
 
@@ -109,14 +119,14 @@
 
 /**
  @abstract Configures the height of each row in the receiver.
- @default 44.0f
+ @default 40.0f
  @raises NSInternalInconsistencyException Raised if the value is mutated after the receiver has been presented.
  */
 @property (nonatomic, assign) CGFloat rowHeight;
 
 /**
  @abstract Configures the sort type of the receiver.
- @default LYRUIParticipantPickerControllerSortTypeFirst
+ @default LYRUIParticipantPickerSortTypeFirstName
  @raises NSInternalInconsistencyException Raised if the value is mutated after the receiver has been presented.
  */
 @property (nonatomic, assign) LYRUIParticipantPickerSortType participantPickerSortType;