+++ Code/LYRUIParticipantTableViewCell.m
@@ -13,10 +13,12 @@
 @interface LYRUIParticipantTableViewCell ()
 
 @property (nonatomic) UILabel *nameLabel;
-@property (nonatomic) UIControl *selectionIndicator;
 @property (nonatomic) LYRUIAvatarImageView *avatarImageView;
+@property (nonatomic) id<LYRUIParticipant> participant;
 @property (nonatomic) LYRUIParticipantPickerSortType sortType;
-@property (nonatomic) BOOL isSelected;
+
+@property (nonatomic) NSLayoutConstraint *nameWithAvatarLeftConstraint;
+@property (nonatomic) NSLayoutConstraint *nameWithoutAvatarLeftConstraint;
 
 @end
 
@@ -26,7 +28,7 @@ static CGFloat const LSSelectionIndicatorSize = 30;
 
 - (id)initWithStyle:(UITableViewCellStyle)style reuseIdentifier:(NSString *)reuseIdentifier
 {
-    self = [super initWithStyle:UITableViewCellStyleSubtitle reuseIdentifier:reuseIdentifier];
+    self = [super initWithStyle:style reuseIdentifier:reuseIdentifier];
     if (self) {
         // UIAppearance Defaults
         _boldTitleFont = [UIFont boldSystemFontOfSize:14];
@@ -34,113 +36,110 @@ static CGFloat const LSSelectionIndicatorSize = 30;
         _titleColor =[UIColor blackColor];
         _subtitleFont = [UIFont systemFontOfSize:12];
         _subtitleColor = [UIColor grayColor];
-        
-        self.textLabel.textAlignment = NSTextAlignmentCenter;
-        self.selectionStyle = UITableViewCellSelectionStyleNone;
-        self.backgroundColor = [UIColor whiteColor];
-        
+
+        self.nameLabel = [UILabel new];
+        self.nameLabel.translatesAutoresizingMaskIntoConstraints = NO;
+        [self.contentView addSubview:self.nameLabel];
+
         self.avatarImageView = [[LYRUIAvatarImageView alloc] init];
         self.avatarImageView.backgroundColor = LYRUILightGrayColor();
         self.avatarImageView.layer.cornerRadius = LSSelectionIndicatorSize / 2;
         self.avatarImageView.translatesAutoresizingMaskIntoConstraints = NO;
-        self.avatarImageView.alpha = 0.0f;
         [self.contentView addSubview:self.avatarImageView];
-        [self updateConstraints];
-        
+
+        [self addConstraint:[NSLayoutConstraint constraintWithItem:self.nameLabel attribute:NSLayoutAttributeTop relatedBy:NSLayoutRelationGreaterThanOrEqual toItem:self.contentView attribute:NSLayoutAttributeTop multiplier:1.0 constant:8]];
+        [self addConstraint:[NSLayoutConstraint constraintWithItem:self.nameLabel attribute:NSLayoutAttributeBottom relatedBy:NSLayoutRelationLessThanOrEqual toItem:self.contentView attribute:NSLayoutAttributeBottom multiplier:1.0 constant:-8]];
+        [self addConstraint:[NSLayoutConstraint constraintWithItem:self.nameLabel attribute:NSLayoutAttributeCenterY relatedBy:NSLayoutRelationEqual toItem:self.contentView attribute:NSLayoutAttributeCenterY multiplier:1.0 constant:0]];
+        [self addConstraint:[NSLayoutConstraint constraintWithItem:self.nameLabel attribute:NSLayoutAttributeRight relatedBy:NSLayoutRelationLessThanOrEqual toItem:self.contentView attribute:NSLayoutAttributeRight multiplier:1.0 constant:-10]];
+        self.nameWithAvatarLeftConstraint = [NSLayoutConstraint constraintWithItem:self.nameLabel attribute:NSLayoutAttributeLeft relatedBy:NSLayoutRelationEqual toItem:self.avatarImageView attribute:NSLayoutAttributeRight multiplier:1.0 constant:15];
+        self.nameWithoutAvatarLeftConstraint = [NSLayoutConstraint constraintWithItem:self.nameLabel attribute:NSLayoutAttributeLeft relatedBy:NSLayoutRelationEqual toItem:self.contentView attribute:NSLayoutAttributeLeft multiplier:1.0 constant:15];
+
+        [self addConstraint:[NSLayoutConstraint constraintWithItem:self.avatarImageView attribute:NSLayoutAttributeLeft relatedBy:NSLayoutRelationEqual toItem:self.contentView attribute:NSLayoutAttributeLeft multiplier:1.0 constant:15]];
+        [self addConstraint:[NSLayoutConstraint constraintWithItem:self.avatarImageView attribute:NSLayoutAttributeCenterY relatedBy:NSLayoutRelationEqual toItem:self.contentView attribute:NSLayoutAttributeCenterY multiplier:1.0 constant:0]];
     }
     return self;
 }
 
-- (void)updateConstraints
+- (void)setHighlighted:(BOOL)highlighted animated:(BOOL)animated
 {
-    [self.contentView addConstraint:[NSLayoutConstraint constraintWithItem:self.avatarImageView attribute:NSLayoutAttributeCenterX relatedBy:NSLayoutRelationEqual toItem:self.imageView attribute:NSLayoutAttributeCenterX multiplier:1.0 constant:0]];
-    [self.contentView addConstraint:[NSLayoutConstraint constraintWithItem:self.avatarImageView attribute:NSLayoutAttributeCenterY relatedBy:NSLayoutRelationEqual toItem:self.imageView attribute:NSLayoutAttributeCenterY multiplier:1.0 constant:0]];
-    [super updateConstraints];
+    // We don't want the default behavior that changes image view backgrounds to transparent while highlighted.
+    UIColor *preservedAvatarBackgroundColor = self.avatarImageView.backgroundColor;
+    [super setHighlighted:highlighted animated:animated];
+    self.avatarImageView.backgroundColor = preservedAvatarBackgroundColor;
 }
 
-- (void)presentParticipant:(id<LYRUIParticipant>)participant
+- (void)setSelected:(BOOL)selected animated:(BOOL)animated
+{
+    // We don't want the default behavior that changes image view backgrounds to transparent while selected.
+    UIColor *preservedAvatarBackgroundColor = self.avatarImageView.backgroundColor;
+    [super setSelected:selected animated:animated];
+    self.avatarImageView.backgroundColor = preservedAvatarBackgroundColor;
+}
+
+- (void)presentParticipant:(id<LYRUIParticipant>)participant withSortType:(LYRUIParticipantPickerSortType)sortType shouldShowAvatarImage:(BOOL)shouldShowAvatarImage
 {
     self.accessibilityLabel = [participant fullName];
-    self.textLabel.text = participant.fullName;
-    [self.avatarImageView setInitialsForName:participant.fullName];
-}
-
-- (void)updateWithSortType:(LYRUIParticipantPickerSortType)sortType
-{
-    _sortType = sortType;
-}
-
-- (void)shouldShowAvatarImage:(BOOL)shouldShowAvatarImage
-{
+    self.participant = participant;
+    self.sortType = sortType;
     if (shouldShowAvatarImage) {
-        self.imageView.backgroundColor = [UIColor redColor];
-        self.imageView.image = [self imageWithColor:[UIColor whiteColor]];
-        [self.imageView addSubview:self.avatarImageView];
-        self.avatarImageView.alpha = 1.0f;
+        [self removeConstraint:self.nameWithoutAvatarLeftConstraint];
+        [self addConstraint:self.nameWithAvatarLeftConstraint];
+        self.avatarImageView.hidden = NO;
+    } else {
+        [self removeConstraint:self.nameWithAvatarLeftConstraint];
+        [self addConstraint:self.nameWithoutAvatarLeftConstraint];
+        self.avatarImageView.hidden = YES;
     }
+    [self.avatarImageView setInitialsForName:participant.fullName];
+    [self configureNameLabel];
 }
 
 - (void)setTitleFont:(UIFont *)titleFont
 {
     _titleFont = titleFont;
+    [self configureNameLabel];
 }
 
 - (void)setBoldTitleFont:(UIFont *)boldTitleFont
 {
     _boldTitleFont = boldTitleFont;
+    [self configureNameLabel];
 }
 
 - (void)setTitleColor:(UIColor *)titleColor
 {
     _titleColor = titleColor;
+    [self configureNameLabel];
 }
 
-- (UIImage *)imageWithColor:(UIColor *)color {
-    CGRect rect = CGRectMake(0, 0, 30, 30);
-    UIGraphicsBeginImageContextWithOptions(rect.size, NO, 0);
-    [color setFill];
-    UIRectFill(rect);   // Fill it with your color
-    UIImage *image = UIGraphicsGetImageFromCurrentImageContext();
-    UIGraphicsEndImageContext();
-    return image;
-}
-
-- (void)layoutSubviews
+- (void)configureNameLabel
 {
-    [super layoutSubviews];
-    
-    if (!self.titleFont) {
+    if (self.participant.fullName.length == 0) {
+        self.nameLabel.text = nil;
         return;
     }
-    
+
+    NSMutableAttributedString *attributedString = [[NSMutableAttributedString alloc] initWithString:self.participant.fullName attributes:@{NSFontAttributeName: self.titleFont}];
+
+    NSRange rangeToBold = NSMakeRange(NSNotFound, 0);
     switch (self.sortType) {
-            
-        case LYRUIParticipantPickerControllerSortTypeFirst: {
-            NSMutableAttributedString *attributedString = [self.textLabel.attributedText mutableCopy];
-            NSRange rangeOfString = [self.textLabel.text rangeOfString:@" "];
-            NSString *regularString = [self.textLabel.text substringFromIndex:rangeOfString.location];
-            NSRange rangeToBold = NSMakeRange(0, rangeOfString.location);
-            [attributedString addAttributes:@{NSFontAttributeName: self.boldTitleFont} range:rangeToBold];
-            [attributedString addAttributes:@{NSFontAttributeName: self.titleFont} range:NSMakeRange(rangeOfString.location, regularString.length)];
-            self.textLabel.attributedText = attributedString;
-        }
+        case LYRUIParticipantPickerSortTypeFirstName:
+            if (self.participant.firstName.length != 0) {
+                rangeToBold = [self.participant.fullName rangeOfString:self.participant.firstName];
+            }
             break;
-            
-        case LYRUIParticipantPickerControllerSortTypeLast: {
-            NSMutableAttributedString *attributedString = [self.textLabel.attributedText mutableCopy];
-            NSRange rangeOfString = [self.textLabel.text rangeOfString:@" "];
-            NSString *stringToBold = [self.textLabel.text substringFromIndex:rangeOfString.location];
-            NSRange rangeToBold = NSMakeRange(rangeOfString.location, stringToBold.length);
-            [attributedString addAttributes:@{NSFontAttributeName: self.titleFont} range:NSMakeRange(0, rangeOfString.location)];
-            [attributedString addAttributes:@{NSFontAttributeName: self.boldTitleFont} range:rangeToBold];
-            self.textLabel.attributedText = attributedString;
-        }
-            break;
-        default:
+        case LYRUIParticipantPickerSortTypeLastName:
+            if (self.participant.lastName.length != 0) {
+                rangeToBold = [self.participant.fullName rangeOfString:self.participant.lastName options:NSBackwardsSearch];
+            }
             break;
     }
-    self.textLabel.textColor = self.titleColor;
-    [self.textLabel sizeToFit];
+    if (rangeToBold.location != NSNotFound) {
+        [attributedString addAttributes:@{NSFontAttributeName: self.boldTitleFont} range:rangeToBold];
+    }
+
+    self.nameLabel.attributedText = attributedString;
+    self.nameLabel.textColor = self.titleColor;
 }
 
 @end